Bugfix: the type of the pathParams was incorrectly set to QueryParam.
Before the type of the urlBuilder was:

```ts
export interface UrlBuilderOptions<QueryParams> {
  url: Url;
  pathParams?: QueryParams;
  queryParams?: QueryParams;
  defaultQueryParams?: QueryParams;
}
```

This is incorrect because `pathParams` is never the same as `queryParams`.

The correct type is:

```ts
export interface UrlBuilderOptions<PathParams, QueryParams> {
  url: Url;
  pathParams?: PathParams;
  queryParams?: QueryParams;
  defaultQueryParams?: QueryParams;
}
```

The PathParams should have its own generic to make this correct.

Appearently this is something that TypeScript `3.5.3` can pick up,
so bumpbed the version to it.

Also wrote a scenario test which makes sure that the TypeScript will
fail in the compile if this bug ever comes back.