Improvement: documentation is now a guide and more detailed.
It also promotes a new way of using the library. It no longer supports
using url functions to do two things: create urls and create meta urls.
It now tells the user to only use the `urlBuilder` to create urls.

Meta urls are now simple `const string` exports which the user should use
in a Route.