fix: booleans from default query params were inverted.
When parsing empty query params with the following defaults:

```ts
function defaultQueryParams() {
  return { visible: true },
};
```

Visible would be `false` and not the expected `true`.

The reason for this is because when the default value was used due to
a missing actual query parameters value, it also got parsed as if it
was a string.

Now whenever we detect that the value is a default value we forgo any
parsing / conversion to an actual value, since that is not needed for
a default value.

In the tests we now test the conversion with an integration test. This
way we test from userspace -> query-params -> mad-react-url. This way
we can catch nasty integration bugs quicker.

Furthermore the wrong `query params` library was used. Instead of
`query-params` the Node.js standard library module `queryparams` was
used.

Also upgraded the queryparams library to the latest version.

Closes #19