fix: multiple optional empty path params didn't strip all ending slashes
Calling `pathParamsBuilder` with multiple missing optional path params
like so:

```ts
const generatedUrl = pathParamsBuilder({
  url: '/users/:id?/:action?/:subaction?',
  pathParams: {},
});
```

Would result in `/users//` and not in the expected `/users`.

The reason for this is because the `stripEndingSlash` would only strip
the final slash, and not all ending slashes.

Renamed `stripEndingSlash` to `stripEndingSlashes` and replaced it with
a regex which removes all ending slashes.

Fixes #21