fix: when urlBuilder was called with '/' an empty string was returned.
When `urlBuilder` was called with the following config:

```ts
urlBuilder({ url: '/' })
```

The result was an empty string instead of the expected `/`.

The reason for this is because the `pathParamsBuilder` will stip out
ending slashes, causing it to return an empty string.

The solution was to make sure that the `stripEndingSlashes` function
does not do anything when the provided url is a '/' (ending slash)
itself.

Also updated all dependencies. By updating TypeScript a
`@ts-expect-error` was no longer needed in `queryParamsBuilder`.

Fixes #31