Bugfix: special chars in queryParams weren't encoded
When a query param had a special char like '&' the char would't be
encode resulting the '&' char to act like the separtor for the next
query param instead of being part of the string.

Reason: the stringify encode option inqueryParamsBuilder was set
to false resulting in no encoding the query params.

Fix: removed the encode false option in the stringify method.

Also updated all dependencies.

Fixes #33