improvement: optional location
When using the browser location, the global window.location is similar
to the location provided by react router for instance. When you need to
use useLocation only to get query params from the current URL, that
makes the code longer and harder to maintain.

Made location optional and default to global window.location.